+++ Cargo.toml
@@ -6,10 +6,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-actix-web = "4"
-diesel = { version = "2", features = ["sqlite", "r2d2"] }
-dotenv = "0.15"
-serde = { version = "1.0", features = ["derive"] }
-uuid = { version = "1", features = ["v4", "serde"] }
-env_logger = "0.10"
+actix-web = "4.4.0"
+diesel = { version = "2.1.4", features = ["sqlite", "r2d2"] }
+dotenv = "0.15.0"
+dotenvy = "0.15.7"
+serde = { version = "1.0.193", features = ["derive"] }
+uuid = { version = "1.6.1", features = ["v4", "serde"] }
+env_logger = "0.10.1"
 log = "0.4"
