+++ src/actions.rs
@@ -1,15 +1,15 @@
 use diesel::{prelude::*, delete};
 use uuid::Uuid;
 
-use crate::models::{self, Task};
+use crate::model::task;
 
 type DbError = Box<dyn std::error::Error + Send + Sync>;
 
 /// Query to get all tasks
-pub fn find_all_tasks(conn: &mut SqliteConnection) -> Result<Vec<models::Task>, DbError> {
+pub fn find_all_tasks(conn: &mut SqliteConnection) -> Result<Vec<task::Task>, DbError> {
     use crate::schema::tasks::dsl::*;
 
-    let get_tasks = tasks.load::<Task>(conn)?;
+    let get_tasks = tasks.load::<task::Task>(conn)?;
     Ok(get_tasks)
 }
 
@@ -17,12 +17,12 @@ pub fn find_all_tasks(conn: &mut SqliteConnection) -> Result<Vec<models::Task>, 
 pub fn find_task_by_uid(
     conn: &mut SqliteConnection,
     uid: Uuid,
-) -> Result<Option<models::Task>, DbError> {
+) -> Result<Option<task::Task>, DbError> {
     use crate::schema::tasks::dsl::*;
 
     let task = tasks
         .filter(id.eq(uid.to_string()))
-        .first::<models::Task>(conn)
+        .first::<task::Task>(conn)
         .optional()?;
 
     Ok(task)
@@ -33,13 +33,13 @@ pub fn insert_new_task(
     conn: &mut SqliteConnection,
     nm: &str,
     dn: &bool,
-) -> Result<models::Task, DbError> {
+) -> Result<task::Task, DbError> {
     // It is common when using Diesel with Actix Web to import schema-related
     // modules inside a function's scope (rather than the normal module's scope)
     // to prevent import collisions and namespace pollution.
     use crate::schema::tasks::dsl::*;
 
-    let new_task = models::Task {
+    let new_task = task::Task {
         id: Uuid::new_v4().to_string(),
         name: nm.to_owned(),
         done: *dn,
