+++ src/main.rs
@@ -1,121 +1,31 @@
-//! Actix Web Diesel integration
-//!
-//! Diesel v2 is not an async library, so we have to execute queries in `web::block` closures which
-//! offload blocking code (like Diesel's) to a thread-pool in order to not block the server.
+use actix_web::{middleware::Logger, web, App, HttpServer};
+use api::tasks::{add_task, delete_task, get_all_tasks, get_task};
 
-#[macro_use]
-extern crate diesel;
-
-use actix_web::web::Json;
-use actix_web::{
-    delete, error, get, middleware, post, web, App, HttpResponse, HttpServer, Responder, Result,
-};
-use serde::Serialize;
-use uuid::Uuid;
-
-use crate::initdb::initialize_db_pool;
-use crate::initdb::DbPool;
-
+mod api;
 mod actions;
-mod initdb;
-mod models;
+mod model;
 mod schema;
-
-#[derive(Serialize)]
-struct Response {
-    message: String,
-}
-
-/// Get all tasks
-#[get("/tasks")]
-async fn get_all_tasks(pool: web::Data<DbPool>) -> Result<impl Responder> {
-    let tasks = web::block(move || {
-        let mut conn = pool.get()?;
-        actions::find_all_tasks(&mut conn)
-    })
-    .await?
-    .map_err(error::ErrorInternalServerError)?;
-
-    Ok(HttpResponse::Ok().json(tasks))
-}
-
-#[delete("/tasks/{task_id}")]
-async fn delete_task(pool: web::Data<DbPool>, task_uid: web::Path<Uuid>,) -> Result<Json<Response>> {
-  let uid: Uuid = task_uid.clone();
-  let conn_result = pool.get();
-
-  match conn_result {
-    Ok(mut conn) => {
-      match actions::destroy_task(&mut conn, uid) {
-        Ok(_rows_deleted) => {
-          Ok(web::Json(Response { message: "deleted".to_string()}))
-        }
-        Err(err) => {
-          Ok(web::Json(Response { message: err.to_string() }))
-        }
-      }
-    }
-    Err(err) => {
-      Ok(web::Json(Response { message: err.to_string() }))
-    }
-  }
-}
-
-/// Finds task by UID.
-#[get("/task/{task_id}")]
-async fn get_task(
-    pool: web::Data<DbPool>,
-    task_uid: web::Path<Uuid>,
-) -> Result<impl Responder> {
-    let task_uid = task_uid.into_inner();
-    let task = web::block(move || {
-        // note that obtaining a connection from the pool is also potentially blocking
-        let mut conn = pool.get()?;
-        actions::find_task_by_uid(&mut conn, task_uid)
-    })
-    .await?
-    // map diesel query errors to a 500 error response
-    .map_err(error::ErrorInternalServerError)?;
-
-    Ok(match task {
-        Some(task) => HttpResponse::Ok().json(task),
-        None => HttpResponse::NotFound().body(format!("No task found with UID: {task_uid}")),
-    })
-}
-
-/// Creates new task.
-#[post("/task")]
-async fn add_task(
-    pool: web::Data<DbPool>,
-    form: web::Json<models::NewTask>,
-) -> Result<impl Responder> {
-    let task = web::block(move || {
-        let mut conn = pool.get()?;
-
-        actions::insert_new_task(&mut conn, &form.name, &form.done)
-    })
-    .await?
-    .map_err(error::ErrorInternalServerError)?;
-
-    Ok(HttpResponse::Created().json(task))
-}
+mod initdb;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    dotenv::dotenv().ok();
-    env_logger::init_from_env(env_logger::Env::new().default_filter_or("info"));
+    dotenvy::dotenv().ok();
+    std::env::set_var("RUST_LOG", "debug");
+    std::env::set_var("RUST_BACKTRACE", "1");
+    env_logger::init();
 
     // initialize DB pool outside of `HttpServer::new` so that it is shared across all workers
-    let pool = initialize_db_pool();
+    let pool = initdb::initialize_db_pool();
 
     log::info!("starting HTTP server at http://localhost:8080");
 
     HttpServer::new(move || {
+        let logger = Logger::default();
         App::new()
             // add DB pool handle to app data; enables use of `web::Data<DbPool>` extractor
             .app_data(web::Data::new(pool.clone()))
             // add request logger middleware
-            .wrap(middleware::Logger::default())
+            .wrap(logger)
             // add route handlers
             .service(get_task)
             .service(add_task)
